Fix ActionDispatch::Request.ip
Since Rack::Request::Helpers is usually included in
ActionDispatch::Request _before_ cloudflare/rails is required,
request.ip is not likely to get patched in the context of a rails
controller.

This checks the most common classes that include the helpers
modules and prepends the trusted proxies module.

I updated the tests to include rspec/rails before cloudflare/rails
in order to ensure ActionDispatch::Request and Rack::Request are
loaded. This more closely matches the autoloading order for a
default rails project.