Add env var check on CLOUDFLARE_PROXY on initialization
We only want the behavior in this gem to kick in if we have our
cloudflare proxy turned on, but don't want to have to uninstall
the gem if we decide to turn it off. As such, check the env var
CLOUDFLARE_PROXY upon initialization- if true, modify the app
such that it excludes cloudflare ips from request.ip and
request.remote_ip. If unset or not true, make no change to standard
rails behavior.