Move where the env var check occurs in railtie
We only want the core behavior to kick in if the env var
CLOUDFLARE_PROXY is true, but performing that check in the
intialize method, and then conditionally prepending very low level
code makes specs in both this gem and in any app that installs
the gem very hard to manage.

Instead, always install the railtie, but have the included overridden
methosds only change behavior if CLOUDFLARE_PROXY is true.