Only trust X-Forwarded-For from the right side (#162)
In #149 we got the extremely helpful `Request#cloudflare?` method which
can assist with checking if a request has been proxied through
Cloudflare. The method checks that there's an unbroken chain of trust
from the `REMOTE_ADDR` (if it exists) through forwarded IPs that only
includes trusted internal proxies and Cloudflare IP addresses.

Each proxy in the chain is supposed to append the IP of the traffic it
received to the *right side* of `X-Forwarded-For`. This means that after
evaluating `REMOTE_ADDR` we should be looking at addresses in
`X-Forwarded-For` starting on the right side, not on the left.

https://github.com/modosc/cloudflare-rails/blob/311875eef57862ecf050fb87415c56ea4823486a/lib/cloudflare_rails/check_trusted_proxies.rb#L23-L25

This ordering means that if a request has the original client as the
first address in `X-Forwarded-For` that we'll incorrectly reject it. It
also means that an attacker could craft a forged `X-Forwarded-For` that
starts with a Cloudflare IP and smuggle in a request that appears to be
valid to the method's logic.

The fix here is to continue to look at `REMOTE_ADDR` first but then to
look at `X-Forwarded-For` in right-to-left order until we run out of
trusted IP addresses (whether internal proxies or Cloudflare IPs).

Fix #161